pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "byoc"
authors = [
  {name = "Kale Kundert", email = "kale@thekunderts.net"},
]
readme = 'README.md'
dynamic = ["version", "description"]
requires-python = "~=3.10"
classifiers = [
  'Programming Language :: Python :: 3',
]
dependencies = [
  'funcy',
  'more_itertools',
  'pipeline_func',
  'smartcall',
  'tidyexc',
]

[project.optional-dependencies]
all = [
  'docopt',
  'mako',
  'nestedtext',
  'pyyaml',
]
test = [
  'pytest',
  'pytest_tmp_files',
  'parametrize_from_file',
  'glom',
  'voluptuous',
]
doc = [
  'sphinx',
  'sphinx_rtd_theme',
  'myst_parser',
  'autoclasstoc',
]

[project.urls]
'Documentation' = 'https://byoc.readthedocs.io/en/latest/'
'Version Control' = 'https://github.com/kalekundert/byoc'
'Bug Tracker' = 'https://github.com/kalekundert/byoc/issues'
'Continuous Integration' = 'https://github.com/kalekundert/byoc/actions'
'Test Coverage' = 'https://app.codecov.io/github/kalekundert/byoc'

[tool.ruff.lint]
select = ["F"]
ignore = ["F405", "F403"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
addopts = "--doctest-modules --doctest-glob='*.rst'"
doctest_optionflags = 'NORMALIZE_WHITESPACE'

[tool.coverage.run]
branch = true
parallel = true
source = ['byoc']

[tool.coverage.report]
exclude_lines = [
    "^\\s*raise AssertionError\\b",
    "^\\s*raise NotImplementedError\\b",
]

[tool.semantic_release]
version_variables = ['byoc/__init__.py:__version__']
build_command = 'python -m pip install build && python -m build'
major_on_zero = false
